truck: move error file location
